git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@705 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-modules/system/trunc/src/system/Math.cpp
@@ -5,7 +5,7 @@ using namespace System;
 
 boolean Math::initializeRandomMachine(const uint64 ui64RandomSeed)
 {
-	srand(ui64RandomSeed);
+	srand((unsigned int)ui64RandomSeed);
 	return true;
 }
